+++ character/tests/forms/test_backgrounds.py
@@ -2,35 +2,12 @@ import pytest
 
 from character.constants.equipment import GearType, ToolType
 from character.forms.backgrounds import (
-    _get_artisans_tools,
     _get_gaming_set_tools,
     _get_holy_symbols,
-    _get_non_spoken_languages,
 )
 from character.models.equipment import GearSettings, ToolSettings
-from character.models.races import Language
-
-from ..factories import CharacterFactory
 
 pytestmark = pytest.mark.django_db
-
-
-def test_get_non_spoken_languages():
-    character = CharacterFactory()
-    if character.species:
-        species_languages = {
-            (language.name, language.get_name_display())
-            for language in character.species.languages.all()
-        }
-    else:
-        species_languages = set()
-    language_choices = _get_non_spoken_languages(character.species)
-    assert species_languages & language_choices == set()
-    all_languages = {
-        (language.name, language.get_name_display())
-        for language in Language.objects.all()
-    }
-    assert language_choices <= all_languages
 
 
 def test_get_holy_symbols():
@@ -47,11 +24,3 @@ def test_get_gaming_set_tools():
         for tool in ToolSettings.objects.filter(tool_type=ToolType.GAMING_SET)
     }
     assert _get_gaming_set_tools() == gaming_set_tools
-
-
-def test_get_artisans_tools():
-    artisans_tools = {
-        (tool.name, tool.get_name_display())
-        for tool in ToolSettings.objects.filter(tool_type=ToolType.ARTISANS_TOOLS)
-    }
-    assert _get_artisans_tools() == artisans_tools